+++ src/app/(pages)/home/_components/home-thread-list/index.tsx
@@ -1,13 +1,14 @@
 'use client';
 
 import { getThreads } from '@/api/threads/get-threads';
-import { HomeThread } from '@/app/(pages)/home/_components/home-thread';
 import { Link, Spinner } from '@/components';
 import { useObserverEffect } from '@/hooks';
 import { css } from '@/styled-system/css';
 import { Flex } from '@/styled-system/jsx';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { Fragment, useRef } from 'react';
+
+import { HomeThread } from './home-thread';
 
 interface ThreadListProps {
   type: 'hot' | 'new';
