+++ src/app/(pages)/home/page.tsx
@@ -1,5 +1,4 @@
-import { HomeThreadList } from '@/app/(pages)/home/_components/home-thread-list';
-import { Sorter } from '@/app/(pages)/home/_components/sorter';
+import { HomeThreadList, Sorter } from '@/app/(pages)/home/_components';
 
 export default function HomePage({ searchParams }: { searchParams: { [key: string]: string | string[] | undefined } }) {
   const mode = searchParams.sort as 'new' | 'hot';
