+++ src/app/(pages)/thread/[id]/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './main-thread';
+export * from './reply-editor';
+export * from './reply-thread-list';
