+++ src/app/(pages)/write/_components/index.tsx
@@ -1,3 +1,2 @@
 export * from './header';
 export * from './write-editor';
-export * from './write-editor-title';
